refactor: CQRS 패턴에 맞게 user 도메인 폴더 구조 및 Command와 Query 영역 분리
변경된 점:
- CQRS 패턴에 맞게 Command영역과 Query 영역으로 폴더 구조 변경 및 파일 분리
