--- src/main/java/com/mtvs/devlinkbackend/oauth2/dto/response/UserClientIndividualListResponseDTO.java
@@ -1,14 +1,0 @@
-package com.mtvs.devlinkbackend.oauth2.dto.response;
-
-import lombok.*;
-
-import java.util.List;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class UserClientIndividualListResponseDTO {
-    private List<UserClientIndividual> data;
-}
