--- src/main/java/com/mtvs/devlinkbackend/oauth2/dto/response/UserClientIndividualSingleResponseDTO.java
@@ -1,12 +1,0 @@
-package com.mtvs.devlinkbackend.oauth2.dto.response;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class UserClientIndividualSingleResponseDTO {
-    private UserClientIndividual data;
-}
