+++ src/main/java/com/mtvs/devlinkbackend/user/command/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package com.mtvs.devlinkbackend.user.command.repository;
+
+import com.mtvs.devlinkbackend.user.command.model.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    void deleteByEpicAccountId(String epicAccountId);
+}
