+++ src/main/java/com/mtvs/devlinkbackend/user/query/model/dto/response/UserClientGroupSingleResponseDTO.java
@@ -1,0 +1,13 @@
+package com.mtvs.devlinkbackend.user.query.model.dto.response;
+
+import com.mtvs.devlinkbackend.user.command.model.entity.Business;
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class UserClientGroupSingleResponseDTO {
+    private Business data;
+}
