+++ src/main/java/com/mtvs/devlinkbackend/user/query/model/dto/response/UserClientIndividualListResponseDTO.java
@@ -1,0 +1,14 @@
+package com.mtvs.devlinkbackend.user.query.model.dto.response;
+
+import lombok.*;
+
+import java.util.List;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class UserClientIndividualListResponseDTO {
+    private List<UserClientIndividual> data;
+}
