+++ src/main/java/com/mtvs/devlinkbackend/user/query/model/dto/response/UserClientIndividualSingleResponseDTO.java
@@ -1,0 +1,12 @@
+package com.mtvs.devlinkbackend.user.query.model.dto.response;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class UserClientIndividualSingleResponseDTO {
+    private UserClientIndividual data;
+}
