+++ src/main/java/com/mtvs/devlinkbackend/user/query/model/dto/response/UserPartnerSingleResponseDTO.java
@@ -1,0 +1,12 @@
+package com.mtvs.devlinkbackend.user.query.model.dto.response;
+
+import com.mtvs.devlinkbackend.user.command.model.entity.Dev;
+import lombok.*;
+
+@Getter @Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class UserPartnerSingleResponseDTO {
+    private Dev data;
+}
