+++ src/main/java/com/mtvs/devlinkbackend/user/query/repository/DevViewRepository.java
@@ -1,0 +1,20 @@
+package com.mtvs.devlinkbackend.user.query.repository;
+
+import com.mtvs.devlinkbackend.user.command.model.entity.Dev;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface DevViewRepository extends JpaRepository<Dev, Long> {
+    Dev findDevByUser_EpicAccountId(String epicAccountId);
+
+    Dev findDevByDevPhone(String devPhone);
+
+    List<Dev> findDevsByDevNameContainingIgnoreCase(String devName);
+
+    List<Dev> findDevsByUser_Nickname(String nickname);
+
+    List<Dev> findDevsByDevEmail(String devEmail);
+}
