+++ src/main/java/com/mtvs/devlinkbackend/user/query/service/EpicBusinessViewService.java
@@ -1,0 +1,4 @@
+package com.mtvs.devlinkbackend.user.query.service;
+
+public class EpicBusinessViewService {
+}
